Configure a specific Bazel version to use
This commit sets up a Bazel WORKSPACE, and adds the script `tools/bazel`
which will invoke a specific version of Bazel for this repository,
downloading it if necessary. This makes it so that every user is
building with the same build system version.

The only prerequisite is python3, and currently only Linux x86_64 is
supported.