Import Bazel rules for ROS 2 (#9)
Pull in github.com/mvukov/rules_ros2. This seems to be the better of the
available options, and also most actively maintained at the moment.
Other options that were considered:
* github.com/ApexAI/rules_ros
* github.com/RobotLocomotion/drake-ros/bazel_ros2_rules

This needs Python 3.11, so pull that in as well. Also set up a .bazelrc
with options that are required for ROS applications to compile (the main
one is C++17).

Write a simple pair of applications (`talker` and `listener`) to show
that this all works. They can be run in separate terminals to see
publishing and subscribing in action.