Add a launch file to launch chatter apps together (#10)
Ros 2's launch system lets us specify what programs to run, where to run
them, what arguments to pass to them, etc. It is also responsible
monitoring the state of the processes launched, and reporting and/or
reacting to changes in the state of those processes.

Add a simple launch file that just launches our `talker` and `listener`
applications, and a Bazel target for it. This lets us spin up both nodes
with the command:
```
bazel run -c opt //src/examples:chatter
```