Add an integration test for talker (#11)
* Add an integration test for talker

This type of test spins up any nodes that you specify and a test node
for each test case.

Change talker's publish rate to use a parameter. Set the parameter to a
specific value for the test, and verify that it indeed publishes at that
frequency. Also verify the published value.